requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
openai>=1.30.0
requests>=2.32.0
beautifulsoup4>=4.12.0
Jinja2>=3.1.0
pytz>=2024.1
supabase>=2.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
firecrawl-py>=1.0.0
matplotlib>=3.8.0
yfinance>=0.2.40
